# ai-specs/specs/api-spec.yml
openapi: 3.1.0
info:
  title: Quermed CRM API
  version: 1.0.0
paths:
...
  /api/v1/accounts/{account_id}/opportunities:
    get:
      tags:
      - accounts
      summary: Opportunities of one account (open first)
      operationId: list_account_opportunities_api_v1_accounts__account_id__opportunities_get
      security:
      - HTTPBearer: []
      parameters:
      - name: account_id
        in: path
        required: true
        schema:
          type: string
          format: uuid
          title: Account Id
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/OpportunitySummaryRead'
                title: Response List Account Opportunities Api V1 Accounts  Account
                  Id  Opportunities Get
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
...
  /api/v1/opportunities:
    get:
      tags:
      - opportunities
      summary: List opportunities (scoped)
      operationId: list_opportunities_api_v1_opportunities_get
      security:
      - HTTPBearer: []
      parameters:
      - name: status
        in: query
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Status
      - name: pipeline_id
        in: query
        required: false
        schema:
          anyOf:
          - type: string
            format: uuid
          - type: 'null'
          title: Pipeline Id
      - name: stage_id
        in: query
        required: false
        schema:
          anyOf:
          - type: string
            format: uuid
          - type: 'null'
          title: Stage Id
      - name: division_id
        in: query
        required: false
        schema:
          anyOf:
          - type: string
            format: uuid
          - type: 'null'
          title: Division Id
      - name: owner_id
        in: query
        required: false
        schema:
          anyOf:
          - type: string
            format: uuid
          - type: 'null'
          title: Owner Id
      - name: account_id
        in: query
        required: false
        schema:
          anyOf:
          - type: string
            format: uuid
          - type: 'null'
          title: Account Id
      - name: is_tender
        in: query
        required: false
        schema:
          anyOf:
          - type: boolean
          - type: 'null'
          title: Is Tender
      - name: is_at_risk
        in: query
        required: false
        schema:
          anyOf:
          - type: boolean
          - type: 'null'
          title: Is At Risk
      - name: close_from
        in: query
        required: false
        schema:
          anyOf:
          - type: string
            format: date
          - type: 'null'
          title: Close From
      - name: close_to
        in: query
        required: false
        schema:
          anyOf:
          - type: string
            format: date
          - type: 'null'
          title: Close To
      - name: q
        in: query
        required: false
        schema:
          anyOf:
          - type: string
            maxLength: 100
          - type: 'null'
          title: Q
      - name: page
        in: query
        required: false
        schema:
          type: integer
          minimum: 1
          default: 1
          title: Page
      - name: page_size
        in: query
        required: false
        schema:
          type: integer
          maximum: 200
          minimum: 1
          default: 50
          title: Page Size
      - name: sort
        in: query
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Sort
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Page_OpportunitySummaryRead_'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
    post:
      tags:
      - opportunities
      summary: Create an opportunity (three fields, smart defaults)
      operationId: create_opportunity_api_v1_opportunities_post
      security:
      - HTTPBearer: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/OpportunityCreate'
      responses:
        '201':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OpportunityRead'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/v1/opportunities/board:
    get:
      tags:
      - opportunities
      summary: Kanban board of one pipeline
      operationId: board_api_v1_opportunities_board_get
      security:
      - HTTPBearer: []
      parameters:
      - name: pipeline_id
        in: query
        required: true
        schema:
          type: string
          format: uuid
          title: Pipeline Id
      - name: division_id
        in: query
        required: false
        schema:
          anyOf:
          - type: string
            format: uuid
          - type: 'null'
          title: Division Id
      - name: owner_id
        in: query
        required: false
        schema:
          anyOf:
          - type: string
            format: uuid
          - type: 'null'
          title: Owner Id
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BoardRead'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/v1/opportunities/{opportunity_id}:
    get:
      tags:
      - opportunities
      summary: Opportunity detail
      operationId: read_opportunity_api_v1_opportunities__opportunity_id__get
      security:
      - HTTPBearer: []
      parameters:
      - name: opportunity_id
        in: path
        required: true
        schema:
          type: string
          format: uuid
          title: Opportunity Id
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OpportunityRead'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
    patch:
      tags:
      - opportunities
      summary: Update descriptive fields
      operationId: update_opportunity_api_v1_opportunities__opportunity_id__patch
      security:
      - HTTPBearer: []
      parameters:
      - name: opportunity_id
        in: path
        required: true
        schema:
          type: string
          format: uuid
          title: Opportunity Id
      - name: If-Match
        in: header
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: If-Match
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/OpportunityUpdate'
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OpportunityRead'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/v1/opportunities/{opportunity_id}/stage:
    post:
      tags:
      - opportunities
      summary: Move to an open stage
      operationId: move_stage_api_v1_opportunities__opportunity_id__stage_post
      security:
      - HTTPBearer: []
      parameters:
      - name: opportunity_id
        in: path
        required: true
        schema:
          type: string
          format: uuid
          title: Opportunity Id
      - name: If-Match
        in: header
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: If-Match
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/StageMove'
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OpportunityRead'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/v1/opportunities/{opportunity_id}/win:
    post:
      tags:
      - opportunities
      summary: Win the opportunity
      operationId: win_opportunity_api_v1_opportunities__opportunity_id__win_post
      security:
      - HTTPBearer: []
      parameters:
      - name: opportunity_id
        in: path
        required: true
        schema:
          type: string
          format: uuid
          title: Opportunity Id
      - name: If-Match
        in: header
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: If-Match
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/OpportunityWin'
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OpportunityRead'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/v1/opportunities/{opportunity_id}/lose:
    post:
      tags:
      - opportunities
      summary: Lose with a reason
      operationId: lose_opportunity_api_v1_opportunities__opportunity_id__lose_post
      security:
      - HTTPBearer: []
      parameters:
      - name: opportunity_id
        in: path
        required: true
        schema:
          type: string
          format: uuid
          title: Opportunity Id
      - name: If-Match
        in: header
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: If-Match
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/OpportunityLose'
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OpportunityRead'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/v1/opportunities/{opportunity_id}/reopen:
    post:
      tags:
      - opportunities
      summary: Reopen a closed opportunity (sales management)
      operationId: reopen_opportunity_api_v1_opportunities__opportunity_id__reopen_post
      security:
      - HTTPBearer: []
      parameters:
      - name: opportunity_id
        in: path
        required: true
        schema:
          type: string
          format: uuid
          title: Opportunity Id
      - name: If-Match
        in: header
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: If-Match
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/OpportunityReopen'
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OpportunityRead'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/v1/opportunities/{opportunity_id}/at-risk:
    post:
      tags:
      - opportunities
      summary: Flag or clear En riesgo (consumables)
      operationId: toggle_at_risk_api_v1_opportunities__opportunity_id__at_risk_post
      security:
      - HTTPBearer: []
      parameters:
      - name: opportunity_id
        in: path
        required: true
        schema:
          type: string
          format: uuid
          title: Opportunity Id
      - name: If-Match
        in: header
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: If-Match
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AtRiskToggle'
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OpportunityRead'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/v1/opportunities/{opportunity_id}/assignment:
    put:
      tags:
      - opportunities
      summary: Reassign the owner (sales management)
      operationId: assign_opportunity_api_v1_opportunities__opportunity_id__assignment_put
      security:
      - HTTPBearer: []
      parameters:
      - name: opportunity_id
        in: path
        required: true
        schema:
          type: string
          format: uuid
          title: Opportunity Id
      - name: If-Match
        in: header
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: If-Match
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/OpportunityAssignment'
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OpportunityRead'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/v1/opportunities/{opportunity_id}/lines:
    post:
      tags:
      - opportunities
      summary: Add a product line
      operationId: add_line_api_v1_opportunities__opportunity_id__lines_post
      security:
      - HTTPBearer: []
      parameters:
      - name: opportunity_id
        in: path
        required: true
        schema:
          type: string
          format: uuid
          title: Opportunity Id
      - name: If-Match
        in: header
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: If-Match
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LineCreate'
      responses:
        '201':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OpportunityRead'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/v1/opportunities/{opportunity_id}/lines/{line_id}:
    patch:
      tags:
      - opportunities
      summary: Update a product line
      operationId: update_line_api_v1_opportunities__opportunity_id__lines__line_id__patch
      security:
      - HTTPBearer: []
      parameters:
      - name: opportunity_id
        in: path
        required: true
        schema:
          type: string
          format: uuid
          title: Opportunity Id
      - name: line_id
        in: path
        required: true
        schema:
          type: string
          format: uuid
          title: Line Id
      - name: If-Match
        in: header
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: If-Match
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LineUpdate'
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OpportunityRead'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
    delete:
      tags:
      - opportunities
      summary: Remove a product line
      operationId: remove_line_api_v1_opportunities__opportunity_id__lines__line_id__delete
      security:
      - HTTPBearer: []
      parameters:
      - name: opportunity_id
        in: path
        required: true
        schema:
          type: string
          format: uuid
          title: Opportunity Id
      - name: line_id
        in: path
        required: true
        schema:
          type: string
          format: uuid
          title: Line Id
      - name: If-Match
        in: header
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: If-Match
      responses:
        '204':
          description: Successful Response
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
...
components:
  schemas:
...
    AtRiskToggle:
      properties:
        flag:
          type: boolean
          title: Flag
      type: object
      required:
      - flag
      title: AtRiskToggle
...
    HTTPValidationError:
      properties:
        detail:
          items:
            $ref: '#/components/schemas/ValidationError'
          type: array
          title: Detail
      type: object
      title: HTTPValidationError
...
    OpportunitySummaryRead:
      properties:
        id:
          type: string
          format: uuid
          title: Id
        account_id:
          type: string
          format: uuid
          title: Account Id
        account_name:
          type: string
          title: Account Name
        name:
          type: string
          title: Name
        pipeline_id:
          type: string
          format: uuid
          title: Pipeline Id
        stage_id:
          type: string
          format: uuid
          title: Stage Id
        stage_name:
          type: string
          title: Stage Name
        division_id:
          type: string
          format: uuid
          title: Division Id
        owner_id:
          type: string
          format: uuid
          title: Owner Id
        owner_name:
          type: string
          title: Owner Name
        status:
          $ref: '#/components/schemas/OpportunityStatus'
        amount:
          type: string
          title: Amount
        expected_close_date:
          type: string
          format: date
          title: Expected Close Date
        is_tender:
          type: boolean
          title: Is Tender
        tender_deadline:
          anyOf:
          - type: string
            format: date
          - type: 'null'
          title: Tender Deadline
        is_at_risk:
          type: boolean
          title: Is At Risk
        stage_entered_at:
          type: string
          format: date-time
          title: Stage Entered At
        days_in_stage:
          type: integer
          title: Days In Stage
        version:
          type: integer
          title: Version
        updated_at:
          anyOf:
          - type: string
            format: date-time
          - type: 'null'
          title: Updated At
      type: object
      required:
      - id
      - account_id
      - account_name
      - name
      - pipeline_id
      - stage_id
      - stage_name
      - division_id
      - owner_id
      - owner_name
      - status
      - amount
      - expected_close_date
      - is_tender
      - tender_deadline
      - is_at_risk
      - stage_entered_at
      - days_in_stage
      - version
      - updated_at
      title: OpportunitySummaryRead
...
    Page_OpportunitySummaryRead_:
      properties:
        items:
          items:
            $ref: '#/components/schemas/OpportunitySummaryRead'
          type: array
          title: Items
        total:
          type: integer
          title: Total
        page:
          type: integer
          title: Page
        page_size:
          type: integer
          title: Page Size
      type: object
      required:
      - items
      - total
      - page
      - page_size
      title: Page[OpportunitySummaryRead]
...
    StageMove:
      properties:
        stage_id:
          type: string
          format: uuid
          title: Stage Id
      type: object
      required:
      - stage_id
      title: StageMove
...
  securitySchemes:
    HTTPBearer:
      type: http
      scheme: bearer